requirements.txt
Django==5.2.6
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0
reportlab==4.0.7
qrcode==7.4.2
PyYAML==6.0.1
python-decouple==3.8
pip install pyotp==2.9.0